Added bitly shortener to journal TOC pages, for issue #1051
